+++ .eslintrc.js
@@ -91,6 +91,13 @@ module.exports = {
         '@typescript-eslint/ban-ts-comment': 'off',
         '@typescript-eslint/no-explicit-any': 'off',
 
+        '@typescript-eslint/no-unused-vars': [
+            'warn',
+            {
+                varsIgnorePattern: '^_'
+            }
+        ],
+
         '@typescript-eslint/naming-convention': [
             'error',
             {