+++ src/lib/managers/CacheManager.ts
@@ -9,8 +9,8 @@ import { typeOf } from '../utils/functions/typeOf.function'
  *
  * Type parameters:
  *
- * - `K` (any) - The cache map key type.
- * - `V` (any) - The cache map value type.
+ * - `K` (`any`) - The cache map key type.
+ * - `V` (`any`) - The cache map value type.
  *
  * @template K The cache map key type.
  * @template V The cache map value type.
@@ -51,7 +51,7 @@ export class CacheManager<K extends string, V> {
      *
      * Type parameters:
      *
-     * - `V` (any) - The type of cache object to return.
+     * - `V` (`any`) - The type of cache object to return.
      *
      * @returns {any} Object representation of the cache map.
      * @template V The type of cache object to return.
@@ -71,7 +71,7 @@ export class CacheManager<K extends string, V> {
      *
      * Type parameters:
      *
-     * - `V` (any) - The type of data being returned.
+     * - `V` (`any`) - The type of data being returned.
      *
      * @param {K} key The key in cache map.
      * @returns {V} The data from cache map.
@@ -121,8 +121,8 @@ export class CacheManager<K extends string, V> {
      *
      * Type parameters:
      *
-     * - `TValue` (any) - The type of data being set.
-     * - `R` (any) - The type of data being returned.
+     * - `TValue` (`any`) - The type of data being set.
+     * - `R` (`any`) - The type of data being returned.
      *
      * @param {K} key The key in cache map.
      * @returns {R} The data from cache map.